added interaction with health and wepons
for giving droids dames use below code
DroidDamge droid = hitObject.GetComponent<DroidDamge>();
droid.TakeDamage(10, false);
// fist is the base damge given to droid 2nd is if it is crit
//crit dame value is worked out by the droid script it just needs to know if it is crit and the base damage

when the droid laser hits the player it will decreases the health of the player

+++ Assets/Scripts/Droids.cs
@@ -14,7 +14,7 @@ public class Droids : MonoBehaviour
     [SerializeField] private float laserSpeed = 10f;
     [SerializeField] private float returnRotationSpeed = 2f;
     [SerializeField] private float helth = 100f;
-    [SerializeField] private float DroidDamage = 10f;
+    [SerializeField] private int DroidDamage = 10;
     [SerializeField] private float critModifier = 2f;
     [SerializeField] private ParticleSystem explosionEffect;
